rename cli/src to numerai/cli for clearer python site-package presence; fix upgrade: init nodes.json before running terraform upgrade
