+++ numerai/cli/node/destroy.py
@@ -1,12 +1,11 @@
 import numerapi
 
-from cli.src.constants import *
-from cli.src.util.files import \
-    format_path_if_mingw,\
+from numerai.cli.constants import *
+from numerai.cli.util.files import \
     load_or_init_nodes,\
     store_config
-from cli.src.util.docker import terraform
-from cli.src.util.keys import get_provider_keys, get_numerai_keys
+from numerai.cli.util.docker import terraform
+from numerai.cli.util.keys import get_provider_keys, get_numerai_keys
 
 
 @click.command()
