+++ numerai/cli/setup.py
@@ -1,10 +1,10 @@
-from cli.src.constants import *
-from cli.src.util.docker import terraform
-from cli.src.util.files import \
+from numerai.cli.constants import *
+from numerai.cli.util.docker import terraform
+from numerai.cli.util.files import \
     maybe_create, \
     format_path_if_mingw, \
     copy_files
-from cli.src.util.keys import \
+from numerai.cli.util.keys import \
     config_numerai_keys, \
     config_provider_keys
