+++ numerai/cli/util/debug.py
